requirement.txt
Flask==3.1.1
pymongo
python-dotenv
flask-cors